refactor: 필요없는 메서드 제거
