+++ backend/src/main/java/com/carffeine/carffeine/station/service/station/StationGridService.java
@@ -1,11 +1,15 @@
 package com.carffeine.carffeine.station.service.station;
 
+import com.carffeine.carffeine.station.domain.station.Coordinate;
 import com.carffeine.carffeine.station.domain.station.Grid;
+import com.carffeine.carffeine.station.domain.station.GridGenerator;
 import com.carffeine.carffeine.station.domain.station.Point;
 import com.carffeine.carffeine.station.infrastructure.repository.station.dto.StationPoint;
+import com.carffeine.carffeine.station.service.station.dto.ClusterRequest;
 import com.carffeine.carffeine.station.service.station.dto.GridWithCount;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Service
@@ -27,13 +31,24 @@ public class StationGridService {
         }
     }
 
-    public List<Grid> assignStationGridsWithCount(List<Grid> grids, List<GridWithCount> gridWithCounts) {
+    public List<Grid> assignStationGridsWithCount(List<GridWithCount> gridWithCounts, ClusterRequest request) {
+        List<Grid> grids = createDisplayGrid(request);
         for (Grid grid : grids) {
             for (GridWithCount gridWithCount : gridWithCounts) {
                 addCountToGrid(grid, gridWithCount);
             }
         }
         return grids;
+    }
+
+    private List<Grid> createDisplayGrid(ClusterRequest request) {
+        GridGenerator gridGenerator = new GridGenerator();
+        Coordinate coordinate = Coordinate.of(request.latitude(), request.latitudeDelta(), request.longitude(), request.longitudeDelta());
+        BigDecimal maxLatitude = coordinate.maxLatitudeValue();
+        BigDecimal minLongitude = coordinate.minLongitudeValue();
+        BigDecimal minLatitude = coordinate.minLatitudeValue();
+        BigDecimal maxLongitude = coordinate.maxLongitudeValue();
+        return gridGenerator.create(Point.of(maxLatitude, minLongitude), Point.of(minLatitude, maxLongitude), request.latitudeDivisionSize(), request.longitudeDivisionSize());
     }
 
     private void addCountToGrid(Grid grid, GridWithCount gridWithCount) {
